requirements.txt
streamlit
google-generativeai
python-dotenv
speechrecognition
reportlab
pillow
gTTS
